% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rank_points.R
\name{calculate_rank_points}
\alias{calculate_rank_points}
\title{Calculate DODV rank points}
\usage{
calculate_rank_points(rank, number_of_boats, regatta_factor)
}
\arguments{
\item{rank}{total rank of a boat in a regatta}

\item{number_of_boats}{}

\item{regatta_factor}{}
}
\value{
DODV rank points
}
\description{
Calculate DODV rank points
}
\examples{
calculate_rank_points(10, 70, 1.2)
calculate_rank_points(3, 30, 1.0)
}
\references{
\href{http://www.dodv.org/regatta/rangliste/ranglisten-regelung/}
}
