% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_regatta_factor.R
\name{calculate_regatta_factor}
\alias{calculate_regatta_factor}
\title{Calculate regatta factor}
\usage{
calculate_regatta_factor(points_median)
}
\arguments{
\item{points_median}{median DODV ranklist points of regatta participants}
}
\value{
calculated regatta factor
}
\description{
Calculate regatta factor
}
\examples{
calculate_regatta_factor(points_median = 60)
calculate_regatta_factor(points_median = 80)
calculate_regatta_factor(points_median = 100)
}
\references{
\url{http://www.dodv.org/regatta/rangliste/erlaeuterung-der-neuen-ranglistenfaktoren-2/}
}
