% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dodv_ranklist.R
\name{click_more}
\alias{click_more}
\title{Helper function: DODV click on 500 button}
\usage{
click_more(sleep_time)
}
\arguments{
\item{sleep_time}{time to wait after click in seconds (to enable site load)}
}
\value{
clicked button
}
\description{
Helper function: DODV click on 500 button
}
