% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dodv_ranklist.R
\name{get_dodv_ranklist}
\alias{get_dodv_ranklist}
\title{Get DODV ranklist}
\usage{
get_dodv_ranklist(url = "https://segler-rangliste.de/dodv/#/opti-a-ger",
  sleep_time = 2, all_years = TRUE, dbg = TRUE)
}
\arguments{
\item{url}{url to a DODV ranklist (default: https://segler-rangliste.de/dodv/#/opti-a-ger)}

\item{sleep_time}{time to wait after click in seconds to enable site load (default: 2)}

\item{all_years}{should ranklists for all years scraped? (default: TRUE)}

\item{dbg}{show debug messages (default: TRUE)}
}
\value{
tibble with ranklists
}
\description{
Get DODV ranklist
}
