% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dodv_regatta_results.R
\name{get_dodv_regatta_overview}
\alias{get_dodv_regatta_overview}
\title{Get DODV regatta overview}
\usage{
get_dodv_regatta_overview(url)
}
\arguments{
\item{url}{url to regatta results table provided by <https://segler-rangliste.de>}
}
\value{
table with regatta results
}
\description{
Get DODV regatta overview
}
