% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dodv_regatta_results.R
\name{get_dodv_regatta_results}
\alias{get_dodv_regatta_results}
\title{Get DODV regatta results}
\usage{
get_dodv_regatta_results(url)
}
\arguments{
\item{url}{url to regatta results table provided by <https://segler-rangliste.de>}
}
\value{
table with regatta results
}
\description{
Get DODV regatta results
}
