% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage2sail.R
\name{manage2sail_registrations}
\alias{manage2sail_registrations}
\title{manage2sail: get regatta registrations}
\usage{

  manage2sail_registrations(url = paste0("https://www.manage2sail.com/de-DE/event/",
  "5b69e6bf-d50a-4536-a203-54e90bd8320d#!/entries?classId=ba596ce5"))
}
\arguments{
\item{url}{url (default: paste0("https://www.manage2sail.com/de-DE/event/", 
"5b69e6bf-d50a-4536-a203-54e90bd8320d#!/entries?classId=ba596ce5", 
"-2827-4b97-8fa4-568ca20659cb"))}
}
\value{
tibble with registrations for regatta
}
\description{
manage2sail: get regatta registrations
}
