% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage2sail.R
\name{manage2sail_results}
\alias{manage2sail_results}
\title{manage2sail: get regatta results}
\usage{
manage2sail_results(url = paste0("http://manage2sail.com/de-DE/event/",
  "117e84c6-177e-4ed2-8432-434b3b362c7e#!/results?classId=437aac44-0937",
  "-4ade-b89f-95161274e6cb"))
}
\arguments{
\item{url}{url (default: paste0("http://manage2sail.com/de-DE/event/", 
"117e84c6-177e-4ed2-8432-434b3b362c7e#!/results?classId=437aac44-0937", 
"-4ade-b89f-95161274e6cb))}
}
\value{
tibble with registrations for regatta
}
\description{
manage2sail: get regatta results
}
