% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raceoffice.R
\name{raceoffice_registrations}
\alias{raceoffice_registrations}
\title{raceoffice: get regatta registrations}
\usage{

  raceoffice_registrations(url = paste0("https://www.manage2sail.com/de-DE/event/",
  "5b69e6bf-d50a-4536-a203-54e90bd8320d#!/entries?classId=ba596ce5-2827",
  "-4b97-8fa4-568ca20659cb"))
}
\arguments{
\item{url}{url (default: paste0("https://www.manage2sail.com/de-DE/event/", 
"5b69e6bf-d50a-4536-a203-54e90bd8320d#!/entries?classId=ba596ce5-2827", 
"-4b97-8fa4-568ca20659cb"))}
}
\value{
tibble with registrations for regatta
}
\description{
raceoffice: get regatta registrations
}
