% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raceoffice.R
\name{raceoffice_registrations}
\alias{raceoffice_registrations}
\title{raceoffice: get regatta registrations}
\usage{
raceoffice_registrations(url = paste0("http://www.raceoffice.org/",
  "entrylist.php?eid=11498618335c794171f0ae7"))
}
\arguments{
\item{url}{url (default: paste0("http://www.raceoffice.org/",
"entrylist.php?eid=11498618335c794171f0ae7"))}
}
\value{
tibble with registrations for regatta
}
\description{
raceoffice: get regatta registrations
}
